utils.py renamed to funcs.py
