+++ utils/main.py
@@ -1,5 +1,5 @@
 import json
-from utils import sorted_dict, client_report
+from funcs import sorted_dict, client_report
 
 '''14.10.2018 Перевод организации
 Visa Platinum 7000 79** **** 6361 -> Счет **9638
